refactor(App): 3개의 함수들을 별도 파일로 분리
