+++ src/printResult.js
@@ -1,0 +1,5 @@
+import { Console } from '@woowacourse/mission-utils';
+
+export function printResult(result) {
+    Console.print(`결과 : ${result}`);
+}
